[#noissue] Fix typo in batch test package
- pinpint -> pinpoint
